Address review: overlay credence/robustness events, mount StagedBanner on stats pages
- Extend compute_project_stats / compute_question_stats RPCs to overlay the
  latest set_credence and set_robustness mutation events for the staged run, so
  credence_histogram and robustness_histogram reflect staged assessments.
- Move StagedBanner to frontend/src/components so both page and stats routes
  share it. Inline its styles so it's drop-in. Mount it (conditional on
  staged_run_id) on the question stats and project stats pages so direct links
  show the banner and sync the useStagedRun context.
- Add test_project_stats_overlays_staged_credence_events to pin the RPC
  overlay behavior.
